+++ src/exceptions.rs
@@ -23,37 +23,13 @@ use std::{
         c_char,
         CStr,
     },
-    sync::{
-        Arc,
-        Mutex,
-        MutexGuard,
-        OnceLock,
-        PoisonError,
-    },
+    sync::Mutex,
 };
 
 use super::QuestError;
 
-#[derive(Default)]
-struct QuestExcept<T: Default>(OnceLock<Arc<Mutex<T>>>);
-
-impl<T> QuestExcept<T>
-where
-    T: Default,
-{
-    const fn new() -> Self {
-        Self(OnceLock::new())
-    }
-
-    fn get_lock(
-        &self
-    ) -> Result<MutexGuard<'_, T>, PoisonError<MutexGuard<'_, T>>> {
-        self.0.get_or_init(Default::default).lock()
-    }
-}
-
-static QUEST_EXCEPT_GUARD: QuestExcept<()> = QuestExcept::new();
-static QUEST_EXCEPT_ERROR: QuestExcept<Vec<QuestError>> = QuestExcept::new();
+static QUEST_EXCEPT_GUARD: Mutex<()> = Mutex::new(());
+static QUEST_EXCEPT_ERROR: Mutex<Vec<QuestError>> = Mutex::new(Vec::new());
 
 /// Report error in a `QuEST` API call.
 ///
@@ -74,7 +50,7 @@ unsafe extern "C" fn invalidQuESTInputError(
     let err_msg = unsafe { CStr::from_ptr(errMsg) }.to_str().unwrap();
     let err_func = unsafe { CStr::from_ptr(errFunc) }.to_str().unwrap();
 
-    QUEST_EXCEPT_ERROR.get_lock().unwrap().insert(
+    QUEST_EXCEPT_ERROR.lock().unwrap().insert(
         0,
         QuestError::InvalidQuESTInputError {
             err_msg:  err_msg.to_owned(),
@@ -102,11 +78,11 @@ where
     F: FnOnce() -> T,
 {
     // Lock QuEST to our call
-    let guard = QUEST_EXCEPT_GUARD.get_lock().unwrap();
+    let guard = QUEST_EXCEPT_GUARD.lock().unwrap();
 
     // The lock here is not bound to any variable; it will be released as
     // soon as the buffer is cleared.
-    QUEST_EXCEPT_ERROR.get_lock().unwrap().clear();
+    QUEST_EXCEPT_ERROR.lock().unwrap().clear();
 
     // Call QuEST API
     let res = f();
@@ -114,7 +90,7 @@ where
     // At this point all exceptions have been thrown.
     // Take the last exception from the buffer (first reported).
     // For now, we log the rest as error messages via invalidQuESTInputError()
-    let err = QUEST_EXCEPT_ERROR.get_lock().unwrap().pop();
+    let err = QUEST_EXCEPT_ERROR.lock().unwrap().pop();
 
     // Drop the guard as soon as we don't need it anymore:
     drop(guard);