+++ src/questenv.rs
@@ -56,3 +56,9 @@ impl Drop for QuestEnv {
             .expect("dropping QuestEnv should always succeed")
     }
 }
+
+// SAFETY:  The way we handle API calls to QuEST by locking the exception
+// handler makes each call atomic and prevent data races.  The promise of this
+// API is that functions that take a shared reference to QuestEnv do not modify
+// the QuEST environment state.
+unsafe impl Sync for QuestEnv {}